:recycle: Refactor
