+++ src/components/UI/Header.tsx
@@ -2,11 +2,7 @@ import { FC, useEffect, useRef, useState } from 'react';
 import autoAnimate from '@formkit/auto-animate';
 
 import useSelectedStore from '../../context/useSelectedStore';
-import { createSharedCollection } from '../../services/todo';
-import useTodoStore from '../../context/useTodoStore';
-import { copyToClipboard } from '../../utils/utils';
 import TodoControls from '../TodoForm/TodoControls';
-import useLanguage from '../../hooks/useLanguage';
 import TodoForm from '../TodoForm/TodoForm';
 import Settings from './Settings/Settings';
 import Confirm from './Confirm';
@@ -24,54 +20,10 @@ const Header: FC = () => {
   const [confirm, setConfirm] = useState<Omit<TConfirm, 'type'> | null>(null);
   const [isLoading, setIsLoading] = useState(false);
   const selectedCollection = useSelectedStore((state) => state.selectedCollection);
-  const { resetSelection } = useSelectedStore((state) => state.actions);
-  const { deleteCollection } = useTodoStore((state) => state.actions);
-  const { text } = useLanguage();
 
   useEffect(() => {
     if (parent.current) autoAnimate(parent.current);
   }, [parent]);
-
-  const deleteConfirmBtnHandler = () => {
-    if (!selectedCollection) return;
-    setIsLoading(true);
-    deleteCollection(selectedCollection.id);
-    resetSelection();
-    setConfirm(null);
-    setIsLoading(false);
-  };
-
-  const shareConfirmBtnHandler = async () => {
-    if (!selectedCollection) return;
-    setIsLoading(true);
-
-    // TODO: useTodoStore
-    await createSharedCollection(selectedCollection);
-    await copyToClipboard(selectedCollection.id);
-    // await editCollection(editedCollection);
-
-    setConfirm(null);
-    setIsLoading(false);
-  };
-
-  const confirmBtnHandler = (confirmType?: TConfirm['type']) => {
-    switch (confirmType) {
-      case 'delete':
-        setConfirm({
-          confirmText: text.controls.deleteConfirm,
-          handler: deleteConfirmBtnHandler,
-        });
-        return;
-      case 'share':
-        setConfirm({
-          confirmText: text.controls.shareConfirm,
-          handler: shareConfirmBtnHandler,
-        });
-        return;
-      default:
-        setConfirm(null);
-    }
-  };
 
   return (
     <header className={styles.header}>
@@ -83,14 +35,14 @@ const Header: FC = () => {
             </h1>
           </div>
           {!selectedCollection && <Settings />}
-          {selectedCollection && <TodoControls onConfirm={confirmBtnHandler} />}
+          {selectedCollection && <TodoControls onConfirm={setConfirm} onLoading={setIsLoading} />}
         </div>
         {!confirm && <TodoForm />}
         {confirm && (
           <Confirm
             confirmText={confirm.confirmText}
             onConfirm={confirm.handler}
-            onCancel={confirmBtnHandler}
+            onCancel={() => setConfirm(null)}
             isLoading={isLoading}
           />
         )}
